requirements.txt
sqlalchemy
astropy
streamlit
pandas
psycopg2-binary
python-dotenv
